Added to_words2 function to convert from number to spanish words
A function to_words2() was created to overcome some limitations and
errors of to_words.

to_words2 can take a number up to 1e22 or a string of digits up to 60
digits long, and convert it to words, thus expanding the capabilities
of to_words.

It also fixes some errors of to_words, particularly the conversion of
quantities ending in "millions", that were converted to "millones y
mil" by to_words().

devtools::document() was run in order to update the documentation of
the package, but some adjustments may be needed in the original
functions, as the following error was displayed:

   ✖ spanish.R:28: `@docType "package"` is deprecated.
   ℹ Please document "_PACKAGE" instead.